fix(web_viewer): switch from QtWebKit to QtWebEngine
As discussed in https://github.com/geotribu/qtribu/pull/99, QtWebKit is deprecated.
The previous fix was not portable. Even when QGIS has QtWebKitWidgets, it fails
if not present on the system. This change adopts a more sustainable method using
only QtWebEngineWidgets, which will also be available in Qt6.

Additionally, instead of calling PyQt (which can be PyQt5 or PyQt6 depending on
the system), we now use the link provided by QGIS with qgis.PyQt.